moving the phantomjs main.js to the puppeteerHandle, making sure the right file is used as well as changing instances of phantom to puppeteer
